Refactor CLI commands and improve server handling
Refactored server and client commands into distinct packages with Cobra integration, enabling flexible command-line usage. Updated server implementations by adding timeouts for HTTP handling and improving error handling. Introduced a Makefile for streamlined testing, linting, and mock generation.
