+++ Makefile
@@ -1,0 +1,8 @@
+test:
+	go test --race ./...
+
+lint:
+	golangci-lint run
+
+mocks:
+	mockery
