+++ pkg/core/clientserver.go
@@ -3,6 +3,7 @@ package core
 import (
 	"context"
 	"net/http"
+	"time"
 
 	"github.com/ksysoev/revdial"
 )
@@ -24,18 +25,22 @@ func (s *ClientServer) Run(ctx context.Context) error {
 	}
 
 	serve := http.Server{
-		Handler: s,
+		Handler:           s,
+		ReadHeaderTimeout: 5 * time.Second,
+		WriteTimeout:      5 * time.Second,
 	}
 
 	go func() {
 		<-ctx.Done()
+
 		_ = serve.Close()
 	}()
 
 	return serve.Serve(listener)
 }
 
-func (s *ClientServer) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+func (s *ClientServer) ServeHTTP(w http.ResponseWriter, _ *http.Request) {
 	w.WriteHeader(http.StatusOK)
-	w.Write([]byte("Hello, World!"))
+
+	_, _ = w.Write([]byte("Hello, World!"))
 }
